+++ cpp/TimelineModel.cpp
@@ -30,6 +30,7 @@
 #include "RoomsModel.h"
 #include "MemberList.h"
 #include "UsersModel.h"
+#include "GlobalObject.h"
 
 Q_DECLARE_METATYPE(QModelIndex)
 
@@ -852,8 +853,8 @@ TimelineModel::cacheMedia(const QString &eventId,
     const auto url  = mxcUrl.toStdString();
     const auto name = QString(mxcUrl).remove(QStringLiteral("mxc://"));
     QFileInfo filename(
-      QStringLiteral("%1/media_cache/%2.%3")
-        .arg(QStandardPaths::writableLocation(QStandardPaths::CacheLocation), name, suffix));
+      QStringLiteral("%1/%2.%3")
+        .arg(GlobalObject::instance()->mediaCachePath(), name, suffix));
     if (QDir::cleanPath(name) != name) {
         nhlog::net()->warn("mxcUrl '{}' is not safe, not downloading file", url);
         return;